move code to src directory
